requirements.txt
azure-functions==1.18.0
pyodbc==5.1.0
SQLAlchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1